fix(readyz): cast extract(epoch) to float8 - numeric panics tokio-postgres
outbox_lag's extract(epoch ...) returns PG numeric, which tokio-postgres
cannot deserialize into Option<f64>; row.get panicked on every /readyz call,
failing the readiness probe (EOF) so the v0.2.0 pod never became ready and the
rollout stuck on the old pod. Validated ::float8 -> double precision against
live PG. (Unit tests miss this - needs the planned PG integration test.)

+++ crates/daimon-pg/src/lib.rs
@@ -83,8 +83,10 @@ impl PgStore {
         let client = self.pool.get().await.ok()?;
         let row = client
             .query_one(
+                // extract(epoch ...) returns `numeric` in PG 14+, which tokio-postgres cannot
+                // deserialize into f64 - cast to float8 explicitly or row.get panics.
                 "SELECT count(*)::bigint AS pending,
-                        extract(epoch FROM now() - min(created_at)) AS oldest_age_secs
+                        extract(epoch FROM now() - min(created_at))::float8 AS oldest_age_secs
                  FROM memory.index_outbox WHERE processed_at IS NULL",
                 &[],
             )
